pyproject.toml
[project]
name = "open-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.6.2",
    "flask>=3.1.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=25.1.0",
    "openai>=2.24.0",
    "pgvector>=0.4.2",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-flask>=1.3.0",
]
